move logger to its own package
This sets the place to centralize all logs configs in its own module
