+++ arthur/__init__.py
@@ -2,17 +2,13 @@
 
 import asyncio
 import os
-from functools import partial
 from typing import TYPE_CHECKING
 
-import loguru
 from pydis_core.utils import apply_monkey_patches
 
 if TYPE_CHECKING:
     from arthur.bot import KingArthur
 
-logger = loguru.logger.opt(colors=True)
-logger.opt = partial(logger.opt, colors=True)
 
 apply_monkey_patches()
