+++ arthur/__main__.py
@@ -9,6 +9,7 @@ from discord.ext import commands
 import arthur
 from arthur.bot import KingArthur
 from arthur.config import CONFIG
+from arthur.log import logger
 
 
 async def main() -> None:
@@ -34,5 +35,5 @@ async def main() -> None:
             await bot.start(CONFIG.token.get_secret_value())
 
 
-with arthur.logger.catch():
+with logger.catch():
     asyncio.run(main())
