+++ arthur/log.py
@@ -1,0 +1,6 @@
+from functools import partial
+
+import loguru
+
+logger = loguru.logger.opt(colors=True)
+logger.opt = partial(logger.opt, colors=True)
